// src/app/feature/Performers/models/Performer.ts
export interface Performer {
  _id?: string;
  name: string;
  performerName: string;
  platformNames: string[];
  location: string;
  performerShift: string;
  performerId: string;
  accountId: string;
  accountType: string;
  bank: string;
  email: string;
  retention: number;
  status: boolean;
}